[BUGFIX] Fix TypeError in PreviewUriBuilder
With #91123 a new PreviewUriBuilder class was introduced.
This class features a factory method `create`, with a type
hinted parameter signature. This now leads to a TypeError
in EXT:workspaces, e.g. when importing datasets in tests.

This is fixed by typecasting the variable to be passed
to the factory method.

Resolves: #93800
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68558
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>

+++ typo3/sysext/workspaces/Classes/Preview/PreviewUriBuilder.php
@@ -175,7 +175,7 @@ class PreviewUriBuilder
         }
 
         // Directly use pid value and consider move placeholders
-        $previewPageId = (empty($versionRecord['pid']) ? $liveRecord['pid'] : $versionRecord['pid']);
+        $previewPageId = (int)(empty($versionRecord['pid']) ? $liveRecord['pid'] : $versionRecord['pid']);
         $additionalParameters = '&previewWS=' . $versionRecord['t3ver_wsid'];
         // Add language parameter if record is a localization
         if (BackendUtility::isTableLocalizable($table)) {
